Fix(vral): fixed a wrong type in vmain register

+++ ppu/vram.go
@@ -27,7 +27,7 @@ func (ppu PPU) getVramAddr() uint16 {
 func (ppu *PPU) vmain(data uint8) uint8 {
 	ppu.vramIncrementMode = data&0x80 != 0
 
-	incrementValues := map[uint8]uint8{
+	incrementValues := map[uint8]uint16{
 		0x00: 1, 0x01: 32, 0x10: 128, 0x11: 128,
 	}
